+++ tests/python/openassetio/managerAPI/test_host.py
@@ -21,18 +21,10 @@ Tests that cover the openassetio.managerAPI.host class.
 # pylint: disable=invalid-name,redefined-outer-name
 # pylint: disable=missing-class-docstring,missing-function-docstring
 
-from unittest import mock
-
 import pytest
 
 from openassetio import Context
-from openassetio.hostAPI import HostInterface
 from openassetio.managerAPI import Host
-
-
-@pytest.fixture
-def mock_host_interface():
-    return mock.create_autospec(spec=HostInterface)
 
 
 @pytest.fixture
@@ -56,16 +48,28 @@ class TestHost():
         assert a_host._interface() is mock_host_interface
 
     def test_identifier(self, host, mock_host_interface):
-        method = mock_host_interface.identifier
-        assert host.identifier() == method.return_value
-        method.assert_called_once_with()
+        expected = "some identifier"
+        mock_host_interface.mock.identifier.return_value = expected
+
+        actual = host.identifier()
+
+        assert actual == expected
+        mock_host_interface.mock.identifier.assert_called_once_with()
 
     def test_displayName(self, host, mock_host_interface):
-        method = mock_host_interface.displayName
-        assert host.displayName() == method.return_value
-        method.assert_called_once_with()
+        expected = "some display name"
+        mock_host_interface.mock.displayName.return_value = expected
+
+        actual = host.displayName()
+
+        assert actual == expected
+        mock_host_interface.mock.displayName.assert_called_once_with()
 
     def test_info(self, host, mock_host_interface):
-        method = mock_host_interface.info
-        assert host.info() == method.return_value
-        method.assert_called_once_with()
+        expected = {"some": "info"}
+        mock_host_interface.mock.info.return_value = expected
+
+        actual = host.info()
+
+        assert actual == expected
+        mock_host_interface.mock.info.assert_called_once_with()