+++ RELEASE_NOTES.md
@@ -42,6 +42,10 @@ v1.0.0-alpha.X
   namespacing macro, i.e. `OPENASSETIO_NS(symbolName)`.
   [#370](https://github.com/OpenAssetIO/OpenAssetIO/issues/370)
 
+- `Host` and `HostInterface` classes have been migrated to C++.
+  Debug and audit functionality is left for future work.
+  [#331](https://github.com/OpenAssetIO/OpenAssetIO/issues/370)
+
 - Switched to preferring un-versioned `clang-tidy` executables when
   the `OPENASSETIO_ENABLE_CLANG_TIDY` build option is enabled. We
   currently target LLVM v12, earlier or later versions may yield